LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-type-layout-invalidation-expected.txt

content

FAIL Changing container-type invalidates layout assert_equals: expected 0 but got 47
